fix: use web worker for importing participants #30
